// 5-lunch-app/src/types/navigationEnums.ts
export enum NavigationItemId {
  MENU = 'menu',
  LUNCH = 'lunch',
  ORDERS = 'orders',
  RATINGS = 'ratings',
}

export enum RoutePath {
  ROOT = '/',
  MENU = '/food-menu',
  LUNCH = '/available-lunch',
  ORDERS = '/your-orders',
  RATINGS = '/ratings',
  LOGIN = '/login',
  REGISTER = '/register',
}

export enum NavigationItemTitle {
  MENU = 'Food Menu',
  LUNCH = 'Available Lunch',
  ORDERS = 'Your Orders',
  RATINGS = 'Ratings',
}